fix: color values became black when missing
This also fixes console log noise when processing certain types of materials.

Closes: #141
