+++ Editor/ShaderSupport/GenericShaderTranslator.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1ab5c137b8744685a5c4809c71150212
+timeCreated: 1752368704
