+++ Editor/ShaderSupport/MaterialExtensions.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c06f96a0d8b04debb153ac4cf2f234bf
+timeCreated: 1752368718
